# app/docs/api.yaml
openapi: 3.0.3
info:
  title: Evora Engine API
  version: 1.0.0
  description: Backend API for the Evora engine platform.
paths:
...
  /api/stop:
    get:
      operationId: stop_list
      description: Returns all stops belonging to the authenticated user's organization.
      summary: List stops
      tags:
      - stop
      security:
      - supabase: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Stop'
          description: ''
        '401':
          description: No JWT or invalid JWT
        '403':
          description: User not associated with any organization
    post:
      operationId: stop_create
      description: Creates a new stop under the authenticated user's organization.
      summary: Create stop
      tags:
      - stop
      security:
      - supabase: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StopCreate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/StopCreate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/StopCreate'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stop'
          description: Stop created
        '400':
          description: Validation error
        '401':
          description: No JWT or invalid JWT
        '403':
          description: User not associated with any organization
...
components:
  schemas:
...
    Stop:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        organization_id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          maxLength: 255
        address:
          type: string
        location:
          type: object
          nullable: true
          properties:
            type:
              type: string
              enum:
              - Point
            coordinates:
              type: array
              items:
                type: number
              minItems: 2
              maxItems: 2
              description: '[longitude, latitude]'
          description: GeoJSON Point (SRID 4326)
        stop_type:
          $ref: '#/components/schemas/StopTypeEnum'
        service_time_minutes:
          type: integer
        parking_capacity:
          type: integer
        workshop_available:
          type: boolean
        fuel_station_available:
          type: boolean
        charging_available:
          type: boolean
        charger_count:
          type: integer
        maintenance_bays:
          type: integer
        operating_hours:
          type: object
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - id
      - name
      - address
      - stop_type
      - updated_at
    StopCreate:
      type: object
      required:
      - name
      - address
      properties:
        name:
          type: string
          maxLength: 255
        address:
          type: string
        location:
          type: object
          nullable: true
          properties:
            type:
              type: string
              enum:
              - Point
            coordinates:
              type: array
              items:
                type: number
              minItems: 2
              maxItems: 2
        stop_type:
          $ref: '#/components/schemas/StopTypeEnum'
        service_time_minutes:
          type: integer
        parking_capacity:
          type: integer
        workshop_available:
          type: boolean
        fuel_station_available:
          type: boolean
        charging_available:
          type: boolean
        charger_count:
          type: integer
        maintenance_bays:
          type: integer
        operating_hours:
          type: object
...
    StopTypeEnum:
      enum:
      - customer
      - warehouse
      - charging_station
      - hub
      type: string
      description: |-
        * `customer` - Customer
        * `warehouse` - Warehouse
        * `charging_station` - Charging Station
        * `hub` - Hub
...
  securitySchemes:
    supabase:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase JWT authentication. Obtain a token by signing in via the
        Supabase client.